+++ resources/views/users/account/parts/shortcut-control.blade.php
